Fix Git case-sensitivity issue for Components and Composition
This commit fixes a Git on Windows case-sensitivity bug where both
uppercase and lowercase directory entries were being tracked:
- Components/__init__.py vs components/__init__.py
- Composition/__init__.py vs composition/__init__.py

The lowercase entries have been removed from Git's index, ensuring
only the correct uppercase versions are tracked.

This completes the package structure refactoring changes to these files.
