+++ src/armodel/models/M2/AUTOSARTemplates/SWComponentTemplate/Components/__init__.py
@@ -1,5 +1,6 @@
 from abc import ABCMeta
 from typing import List
+from .....M2.AUTOSARTemplates.SWComponentTemplate.SwComponentType import SwComponentType
 from .....M2.AUTOSARTemplates.SWComponentTemplate.Composition import AssemblySwConnector, DelegationSwConnector, SwComponentPrototype, SwConnector
 from .....M2.AUTOSARTemplates.SWComponentTemplate.Components.InstanceRefs import InnerPortGroupInCompositionInstanceRef
 from .....M2.AUTOSARTemplates.SWComponentTemplate.SwcInternalBehavior import SwcInternalBehavior
@@ -249,56 +250,6 @@ class PortGroup(Identifiable):
         return self._outer_port_ref
 
 
-class SwComponentType(ARElement, metaclass=ABCMeta):
-    def __init__(self, parent: ARObject, short_name: str):
-        super().__init__(parent, short_name)
-
-        self.ports = []                                                     # type: List[PortPrototype]
-        self.portGroups = []                                                # type: List[PortGroup]
-
-    def getPorts(self):
-        return self.ports
-
-    def createPPortPrototype(self, short_name: str) -> PPortPrototype:
-        prototype = PPortPrototype(self, short_name)
-        self.addElement(prototype)
-        self.ports.append(prototype)
-        return prototype
-
-    def createRPortPrototype(self, short_name) -> RPortPrototype:
-        prototype = RPortPrototype(self, short_name)
-        self.addElement(prototype)
-        self.ports.append(prototype)
-        return prototype
-    
-    def createPRPortPrototype(self, short_name) -> PRPortPrototype:
-        prototype = PRPortPrototype(self, short_name)
-        self.addElement(prototype)
-        self.ports.append(prototype)
-        return prototype
-    
-    def getPPortPrototypes(self) -> List[PPortPrototype]:
-        return list(sorted(filter(lambda c: isinstance(c, PPortPrototype), self.ports), key=lambda o: o.short_name))
-
-    def getRPortPrototypes(self) -> List[RPortPrototype]:
-        return list(sorted(filter(lambda c: isinstance(c, RPortPrototype), self.ports), key=lambda o: o.short_name))
-    
-    def getPRPortPrototypes(self) -> List[PRPortPrototype]:
-        return list(sorted(filter(lambda c: isinstance(c, PRPortPrototype), self.ports), key=lambda o: o.short_name))
-    
-    def getPortPrototypes(self) -> List[PortPrototype]:
-        return list(sorted(filter(lambda c: isinstance(c, PortPrototype), self.ports), key=lambda o: o.short_name))
-
-    def getPortGroups(self) -> List[PortGroup]:
-        return self.portGroups
-
-    def createPortGroup(self, short_name) -> PortGroup:
-        port_group = PortGroup(self, short_name)
-        self.addElement(port_group)
-        self.portGroups.append(port_group)
-        return port_group
-    
-
 class AtomicSwComponentType(SwComponentType, metaclass=ABCMeta):
     def __init__(self, parent: ARObject, short_name: str):
         super().__init__(parent, short_name)
@@ -403,61 +354,3 @@ class ServiceProxySwComponentType(AtomicSwComponentType):
 class ServiceSwComponentType(AtomicSwComponentType):
     def __init__(self, parent: ARObject, short_name: str):
         super().__init__(parent, short_name)
-
-
-class CompositionSwComponentType(SwComponentType):
-    def __init__(self, parent: ARObject, short_name: str):
-        super().__init__(parent, short_name)
-
-        self.components = []                                    # type: List[SwComponentPrototype]
-        self.constantValueMappingRefs = []                      # type: List[RefType]
-        self.dataTypeMappingRefs = []                           # type: List[RefType]
-        self.instantiationRTEEventProps = []                    # type: List[InstantiationRTEEventProps]
-
-    def removeAllAssemblySwConnector(self):
-        for sw_connector in self.getAssemblySwConnectors():
-            self.elements.remove(sw_connector)
-
-    def removeAllDelegationSwConnector(self):
-        for sw_connector in self.getDelegationSwConnectors():
-            self.elements.remove(sw_connector)
-
-    def createAssemblySwConnector(self, short_name: str) -> AssemblySwConnector:
-        if not self.IsElementExists(short_name):
-            connector = AssemblySwConnector(self, short_name)
-            self.addElement(connector)
-        return self.getElement(short_name, AssemblySwConnector)
-
-    def createDelegationSwConnector(self, short_name: str) -> DelegationSwConnector:
-        if not self.IsElementExists(short_name):
-            connector = DelegationSwConnector(self, short_name)
-            self.addElement(connector)
-        return self.getElement(short_name, DelegationSwConnector)
-
-    def getAssemblySwConnectors(self) -> List[AssemblySwConnector]:
-        return list(sorted(filter(lambda e: isinstance(e, AssemblySwConnector), self.elements), key=lambda c: c.short_name))
-
-    def getDelegationSwConnectors(self) -> List[DelegationSwConnector]:
-        return list(sorted(filter(lambda e: isinstance(e, DelegationSwConnector), self.elements), key=lambda c: c.short_name))
-
-    # def getSwConnectors(self) -> List[SwConnector]:
-    #    return list(sorted(filter(lambda e: isinstance(e, SwConnector), self.elements), key=lambda c: c.short_name))
-
-    def getSwConnectors(self) -> List[SwConnector]:
-        return list(filter(lambda e: isinstance(e, SwConnector), self.elements))
-
-    def createSwComponentPrototype(self, short_name: str) -> SwComponentPrototype:
-        if not self.IsElementExists(short_name):
-            prototype = SwComponentPrototype(self, short_name)
-            self.addElement(prototype)
-            self.components.append(prototype)
-        return self.getElement(short_name, SwComponentPrototype)
-
-    def getComponents(self) -> List[SwComponentPrototype]:
-        return self.components
-
-    def addDataTypeMapping(self, data_type_mapping_ref: RefType):
-        self.dataTypeMappingRefs.append(data_type_mapping_ref)
-
-    def getDataTypeMappings(self) -> List[RefType]:
-        return self.dataTypeMappingRefs
