+++ src/armodel/models/M2/AUTOSARTemplates/SWComponentTemplate/Composition/__init__.py
@@ -1,6 +1,8 @@
 from abc import ABCMeta
+from typing import List
 from .....M2.AUTOSARTemplates.SWComponentTemplate.Composition.InstanceRefs import PPortInCompositionInstanceRef, PortInCompositionTypeInstanceRef
 from .....M2.AUTOSARTemplates.SWComponentTemplate.Composition.InstanceRefs import RPortInCompositionInstanceRef
+from .....M2.AUTOSARTemplates.SWComponentTemplate.SwComponentType import SwComponentType
 from .....M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import RefType
 from .....M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ArObject import ARObject
 from .....M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Identifiable import Identifiable
@@ -98,3 +100,58 @@ class PassThroughSwConnector(SwConnector):
     def setRequiredOuterPortRef(self, value: RefType):
         self.requiredOuterPortRef = value
         return self
+
+
+class CompositionSwComponentType(SwComponentType):
+    def __init__(self, parent: ARObject, short_name: str):
+        super().__init__(parent, short_name)
+
+        self.components = []                                    # type: List[SwComponentPrototype]
+        self.constantValueMappingRefs = []                      # type: List[RefType]
+        self.dataTypeMappingRefs = []                           # type: List[RefType]
+        self.instantiationRTEEventProps = []                    # type: List[InstantiationRTEEventProps]
+
+    def removeAllAssemblySwConnector(self):
+        for sw_connector in self.getAssemblySwConnectors():
+            self.elements.remove(sw_connector)
+
+    def removeAllDelegationSwConnector(self):
+        for sw_connector in self.getDelegationSwConnectors():
+            self.elements.remove(sw_connector)
+
+    def createAssemblySwConnector(self, short_name: str) -> AssemblySwConnector:
+        if not self.IsElementExists(short_name):
+            connector = AssemblySwConnector(self, short_name)
+            self.addElement(connector)
+        return self.getElement(short_name, AssemblySwConnector)
+
+    def createDelegationSwConnector(self, short_name: str) -> DelegationSwConnector:
+        if not self.IsElementExists(short_name):
+            connector = DelegationSwConnector(self, short_name)
+            self.addElement(connector)
+        return self.getElement(short_name, DelegationSwConnector)
+
+    def getAssemblySwConnectors(self) -> List[AssemblySwConnector]:
+        return list(sorted(filter(lambda e: isinstance(e, AssemblySwConnector), self.elements), key=lambda c: c.short_name))
+
+    def getDelegationSwConnectors(self) -> List[DelegationSwConnector]:
+        return list(sorted(filter(lambda e: isinstance(e, DelegationSwConnector), self.elements), key=lambda c: c.short_name))
+
+    def getSwConnectors(self) -> List[SwConnector]:
+        return list(filter(lambda e: isinstance(e, SwConnector), self.elements))
+
+    def createSwComponentPrototype(self, short_name: str) -> SwComponentPrototype:
+        if not self.IsElementExists(short_name):
+            prototype = SwComponentPrototype(self, short_name)
+            self.addElement(prototype)
+            self.components.append(prototype)
+        return self.getElement(short_name, SwComponentPrototype)
+
+    def getComponents(self) -> List[SwComponentPrototype]:
+        return self.components
+
+    def addDataTypeMapping(self, data_type_mapping_ref: RefType):
+        self.dataTypeMappingRefs.append(data_type_mapping_ref)
+
+    def getDataTypeMappings(self) -> List[RefType]:
+        return self.dataTypeMappingRefs
